# docs/openapi.yaml
openapi: 3.1.1
info:
  title: AgentFeedback API
  version: '1.0'
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: >-
    HTTP API of AgentFeedback, a service that stores write-once feedback from AI
    coding agents. The contract is this document together with the JSON Schemas
    under `schemas/` and the conformance kit under `conformance/`. This is the API
    of the next major release. Schemas use the custom keyword `x-max-bytes` for
    UTF-8 byte limits, since JSON Schema `maxLength` counts code points. The
    server serves this document at `/api/v1/openapi.json` in bundled form, with
    the referenced schema files inlined.
servers:
  - url: https://api.agentfeedback.io
    description: Hosted
  - url: http://127.0.0.1:8090
    description: Local
security:
  - bearer: []
  - apiKey: []
...
paths:
  /api/v1/submissions:
    post:
      tags: [submissions]
      operationId: createSubmission
      summary: Create a submission
      description: >-
        Accepts any JSON object. The envelope schema describes what is stored
        after inference and is never used to reject a body; deviations are
        normalised and reported as `warnings`. Returns 201 with `Location` for a
        new record, or 200 with `Location` of the existing record for a keyed
        replay of identical content or a keyless duplicate. A body that is not
        JSON or not an object is 400 `bad_request`. A reused `key` with
        different content is 409 `replay_mismatch`. A body over
        `limits.body_bytes` is 413 `request_too_large`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/envelope.v1.json'
      responses:
        '201':
          description: Created.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateResponse'}
              example:
                submission:
                  id: 44
                  uid: 019250f2-8c3e-7a4b-9d1e-2f3a4b5c6d7e
                  kind: friction
                  schema_version: 1
                  key: friction-workstation-a-019250f2
                  summary: README install step 3 references a flag that no longer exists
                  machine: workstation-a
                  model: claude-fable-5-1
                  harness: claude-code
                  project: example
                  occurred_at: '2026-09-27T09:58:12.000000Z'
                  context: {git_commit: a1b2c3d, client: agentfeedback/4.0.0}
                  payload: {category: documentation, details: The --fast flag was removed in 3.0., fix_status: applied, fix_ref: example@a1b2c3d}
                  content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                  created_at: '2026-09-27T10:00:01.123456Z'
                warnings: []
        '200':
          description: Replay of an existing record (keyed replay or keyless duplicate).
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
            Location: {$ref: '#/components/headers/Location'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CreateResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    get:
      tags: [submissions]
      operationId: listSubmissions
      summary: List submissions
      description: >-
        All parameters are optional. Unknown parameter names, empty values and
        repeated singleton parameters are 400 `validation_error` naming the
        parameter. Rows are the record minus `payload` (present with
        `include=payload`). `total` counts every row matching the filters,
        ignoring cursors and `limit`, in the same read transaction. When
        `has_more` is true exactly one of `next_before_id` / `next_after_id` is
        set, matching the direction used.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - $ref: '#/components/parameters/SchemaVersion'
        - $ref: '#/components/parameters/Key'
        - $ref: '#/components/parameters/Machine'
        - $ref: '#/components/parameters/Model'
        - $ref: '#/components/parameters/Project'
        - $ref: '#/components/parameters/Harness'
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/FixStatus'
        - $ref: '#/components/parameters/ExcludeKind'
        - $ref: '#/components/parameters/Verdict'
        - $ref: '#/components/parameters/Processed'
        - $ref: '#/components/parameters/Redacted'
        - $ref: '#/components/parameters/ContentHash'
        - $ref: '#/components/parameters/Since'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/On'
        - $ref: '#/components/parameters/Q'
        - name: before_id
          in: query
          description: Newest-first page of rows with `id < before_id`. Mutually exclusive with `after_id`.
          schema: {type: integer, minimum: 1}
        - name: after_id
          in: query
          description: Oldest-first page of rows with `id > after_id`. Mutually exclusive with `before_id`.
          schema: {type: integer, minimum: 0}
        - name: limit
          in: query
          description: 1–500, or 1–100 with `include=payload`; default 50. Out of range is 400.
          schema: {type: integer, minimum: 1, maximum: 500, default: 50}
        - name: include
          in: query
          description: '`payload`: rows carry `payload`.'
          schema: {type: string, enum: [payload]}
      responses:
        '200':
          description: A page of submissions.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListResponse'}
              example:
                submissions:
                  - id: 1188
                    uid: 019250f2-8c3e-7a4b-9d1e-2f3a4b5c6d7e
                    kind: friction
                    schema_version: 1
                    summary: README install step 3 references a flag that no longer exists
                    machine: workstation-a
                    project: example
                    content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                    created_at: '2026-09-27T10:00:01.123456Z'
                limit: 50
                total: 245
                has_more: true
                next_before_id: 1187
                next_after_id: null
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/submissions/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [submissions]
      operationId: getSubmission
      summary: Get one submission
      description: >-
        Returns the bare record. `{id}` must match `^[1-9][0-9]*$`; anything
        else is 400. 404 `not_found` when absent.
      responses:
        '200':
          description: The record.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Submission'}
              example:
                id: 44
                uid: 019250f2-8c3e-7a4b-9d1e-2f3a4b5c6d7e
                kind: friction
                schema_version: 1
                key: friction-workstation-a-019250f2
                summary: README install step 3 references a flag that no longer exists
                machine: workstation-a
                model: claude-fable-5-1
                harness: claude-code
                project: example
                occurred_at: '2026-09-27T09:58:12.000000Z'
                context: {git_commit: a1b2c3d, client: agentfeedback/4.0.0}
                payload: {category: documentation, details: The --fast flag was removed in 3.0., fix_status: applied, fix_ref: example@a1b2c3d}
                content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                created_at: '2026-09-27T10:00:01.123456Z'
                processed_at: '2026-09-27T14:02:00.000000Z'
                verdict: fixed
                resolution: already applied by the reporter
                ref: example@a1b2c3d
                processed_by: workstation-b/triage-2026-09-27
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    patch:
      tags: [processing]
      operationId: markSubmission
      summary: Set the processing mark of one submission
      description: >-
        Validated strictly (400 `validation_error`). Marking sets
        `processed_at` if unset and stores the given fields; re-marking with a
        different `verdict`/`resolution`/`ref` replaces them and keeps the
        original `processed_at`; `processed=false` clears all five processing
        fields. Redacted rows can still be marked. No content field is
        writable. Returns the record; 404 when absent.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProcessingMark'}
            example:
              processed: true
              verdict: fixed
              resolution: already applied by the reporter
              ref: example@1a2b3c4
              processed_by: workstation-b/session-7
      responses:
        '200':
          description: The updated record.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Submission'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    delete:
      tags: [submissions]
      operationId: redactSubmission
      summary: Redact a submission (tombstone)
      description: >-
        Replaces `payload` with `{"redacted": true}`, removes `summary` and
        `context`, sets `redacted_at`; keeps `id`, `uid`, `kind`,
        `schema_version`, `key`, `machine`, `model`, `harness`, `project`,
        `occurred_at`, `content_hash`, `created_at` and all processing fields.
        Returns 200 with the tombstone; repeating it is 200 unchanged; 404 when
        absent. A keyed replay of the original content still returns 200 with
        the tombstone because the hash is kept. Redaction is not reversible.
      responses:
        '200':
          description: The tombstone.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Submission'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/submissions/processed:
    post:
      tags: [processing]
      operationId: markSubmissions
      summary: Set the processing mark of a batch
      description: >-
        Same fields and rules as the per-row mark, plus `ids` (1–500 positive
        integers, duplicates collapsed). Returns a classification of the ids
        into `updated`, `unchanged` and `not_found`. Validated strictly (400
        `validation_error`).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BatchMark'}
            example:
              ids: [43, 44, 999]
              processed: true
              verdict: fixed
      responses:
        '200':
          description: Classification of the ids.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BatchMarkResponse'}
              example:
                processed: true
                verdict: fixed
                updated: [43]
                unchanged: [44]
                not_found: [999]
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/stats:
    get:
      tags: [stats]
      operationId: getStats
      summary: Aggregates
      description: >-
        Aggregates over the same filters as the list route (everything except
        the cursors, `limit` and `include`). Groups are sorted by `total`
        descending, at most 100. Invalid parameters are 400
        `validation_error`.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - $ref: '#/components/parameters/SchemaVersion'
        - $ref: '#/components/parameters/Key'
        - $ref: '#/components/parameters/Machine'
        - $ref: '#/components/parameters/Model'
        - $ref: '#/components/parameters/Project'
        - $ref: '#/components/parameters/Harness'
        - $ref: '#/components/parameters/Category'
        - $ref: '#/components/parameters/FixStatus'
        - $ref: '#/components/parameters/ExcludeKind'
        - $ref: '#/components/parameters/Verdict'
        - $ref: '#/components/parameters/Processed'
        - $ref: '#/components/parameters/Redacted'
        - $ref: '#/components/parameters/ContentHash'
        - $ref: '#/components/parameters/Since'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/On'
        - $ref: '#/components/parameters/Q'
        - name: by
          in: query
          description: >-
            Comma list of up to three of `kind`, `project`, `category`,
            `fix_status`, `machine`, `model`, `harness`, `verdict`,
            `schema_version`. Absent: no groups. A row without a value for
            any listed key is in no group, so group totals need not add up
            to `total`.
          schema: {type: string}
        - name: top
          in: query
          description: 0–50, default 10; how many recurring hashes to return.
          schema: {type: integer, minimum: 0, maximum: 50, default: 10}
        - name: bucket
          in: query
          description: When set, adds a time series on the `on` timestamp.
          schema: {type: string, enum: [day, week]}
      responses:
        '200':
          description: Aggregates.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Stats'}
              example:
                total: 245
                open: 61
                processed: 184
                redacted: 2
                groups:
                  - keys: {project: example, category: documentation}
                    total: 40
                    open: 12
                    processed: 28
                recurring:
                  - content_hash: 9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a9f3a
                    count: 7
                    first_id: 1101
                    last_id: 1187
                    summary: README step 3 flag gone
                    kind: friction
                    project: example
                series:
                  - bucket: '2026-09-22'
                    total: 31
                    open: 4
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/export:
    get:
      tags: [transfer]
      operationId: exportSubmissions
      summary: Export as NDJSON (format 2)
      description: >-
        One read transaction, ascending `id`. The body is a header line, one
        line per full record (tombstones exported as tombstones), and a trailer
        line. `sha256` in the trailer is over the record lines only, each with
        its trailing newline, header and trailer excluded, in order.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - name: since
          in: query
          description: RFC 3339, inclusive, on `created_at`; `+` must be percent-encoded.
          schema: {type: string, format: date-time}
        - name: after_id
          in: query
          description: Only rows with `id > after_id`.
          schema: {type: integer, minimum: 0}
        - name: limit
          in: query
          description: 1–500; absent means everything.
          schema: {type: integer, minimum: 1, maximum: 500}
      responses:
        '200':
          $ref: '#/components/responses/Export'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
    head:
      tags: [transfer]
      operationId: exportSubmissionsHead
      summary: Export headers only
      description: Same as `GET /api/v1/export`, headers only.
      parameters:
        - $ref: '#/components/parameters/Kind'
        - name: since
          in: query
          description: RFC 3339, inclusive, on `created_at`.
          schema: {type: string, format: date-time}
        - name: after_id
          in: query
          description: Only rows with `id > after_id`.
          schema: {type: integer, minimum: 0}
        - name: limit
          in: query
          description: 1–500; absent means everything.
          schema: {type: integer, minimum: 1, maximum: 500}
      responses:
        '200':
          description: Headers of the export response.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Cache-Control: {$ref: '#/components/headers/CacheControl'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /api/v1/schemas:
    get:
      tags: [schemas]
      operationId: listSchemas
      summary: List known schemas
      description: Kinds and versions the server knows, including `envelope`. Unauthenticated; carries no data.
      security: []
      responses:
        '200':
          description: Known schemas.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SchemasList'}
              example:
                schemas:
                  - {kind: envelope, versions: [1]}
                  - {kind: friction, versions: [1]}
                  - {kind: review, versions: [1]}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/schemas/{kind}/{version}:
    get:
      tags: [schemas]
      operationId: getSchema
      summary: Get one schema
      description: >-
        Returns the JSON Schema document verbatim with `Content-Type:
        application/schema+json`. `envelope` is a valid `{kind}`. 404 when the
        pair is unknown. Unauthenticated; carries no data.
      security: []
      parameters:
        - name: kind
          in: path
          required: true
          description: Kind token, or `envelope`.
          schema: {type: string}
        - name: version
          in: path
          required: true
          description: Schema version.
          schema: {type: integer, minimum: 1}
      responses:
        '200':
          description: The JSON Schema document.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/schema+json:
              schema: {type: object}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
...
  /api/v1/openapi.json:
    get:
      tags: [discovery]
      operationId: getOpenApi
      summary: This OpenAPI document
      description: The OpenAPI document as JSON. Unauthenticated; carries no data.
      security: []
      responses:
        '200':
          description: The OpenAPI document.
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {type: object}
        '405': {$ref: '#/components/responses/MethodNotAllowed'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <key>`; scheme name case-insensitive. Either header authorises; neither shadows the other.'
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: '`X-Api-Key: <key>`. Either header authorises; neither shadows the other.'
  headers:
    XRequestId:
      description: Request identifier; present on every response.
      schema: {type: string}
    CacheControl:
      description: '`no-store` on every response of an authenticated route, success or error.'
      schema: {type: string, const: no-store}
...
    Location:
      description: Path of the stored record, e.g. `/api/v1/submissions/44`.
      schema: {type: string}
...
  parameters:
...
    Key:
      name: key
      in: query
      description: Exact key.
      schema: {type: string}
...